refactor(tests): Move constants to tests.constants
