--- tests/__init__.py
@@ -1,11 +1,0 @@
-"""Tests for tmuxp.
-
-tmuxp.tests
-~~~~~~~~~~~
-
-"""
-import pathlib
-
-TESTS_PATH = pathlib.Path(__file__).parent
-EXAMPLE_PATH = TESTS_PATH.parent / "examples"
-FIXTURES_PATH = TESTS_PATH / "fixtures"
