+++ tests/constants.py
@@ -1,0 +1,5 @@
+import pathlib
+
+TESTS_PATH = pathlib.Path(__file__).parent
+EXAMPLE_PATH = TESTS_PATH.parent / "examples"
+FIXTURE_PATH = TESTS_PATH / "fixtures"
